knoogle-server/requirements.txt
AdvancedHTMLParser==8.1.8
autopep8==1.4.4
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.10
django-allauth==0.40.0
django-cors-headers==3.1.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
fake-useragent==0.1.11
gunicorn==19.9.0
idna==2.8
oauthlib==3.1.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
PyPDF2==1.26.0
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.3
QueryableList==3.1.0
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
